Make it possible to use boot.iso again for kickstart tests
A change in virt-install means that if we want to continue using
boot ISOs to run kickstart tests, we need to specify where kernel
and initrd are located on the boot ISO.

This is not needed for full DVD images, but those are much bigger
and also fails to compose much more often than boot ISOs for Rawhide.

So lets specify the kernel and initrd paths to virt-install so
that we can continue using boot ISOs for kickstart tests.