Change default boot options
Drop the default option forcing the "vnc" UI mode.
This was overriding the NTP tests that try to
run in text mode and also is pretty much useless
with tests running on remote workers.

Also add inst.noninteractive option so that we don't
wait hours for tests that get stuck blocking on of the
"channels" for some reason, quite likely unrelated to
what's being tested.