update network noipv4 tests for kickstart applied in stage 2
With processing of kickstart in stage 2 the code path of ordinary kickstart is
almost the same as for the configuration in %pre section, therefore the -httpks
and -pre versions of the test are converging to do the same checks. In the
future, when we are sure processing kickstart in stage2 has no fatal drawbacks
I am probably going to remove (or reduce to a few cases) the number of -pre
variants of networking tests.

Add a device with both --noipv4 and --noipv6 - in this case the connection is
activated.

Fix the case with --noipv4 only - in this case the connection is not activated
and is replaced by default autoconnection (on systems where default
autoconnections are enabled).