Revert "Define network names only in defaults file"
This reverts commit 117b5e2c1fa47a1056db5b0ccd73171aba2cdbe6.

The variables need to be available also in .sh test files which was nat the
case for running tests of remote hosts as the defaults file was sourced in
run_kicsktart_tests.sh and the variables were not passed by parallel to remote
host.

It would be possible to pass the variables via --env option but they would have
to be enumerated (--env _ seems not to work). Using single cfg file seems
better that enumerating the variables at two places.